learning more about stuff, hooked up to context using reducers
different sections have encapsulated functionality based on useContext paired with useReducer. Context state of "section" is now available everywhere for import. currently imported in Content for setting based on scroll position, and Background to render differently based on the current  section.
this is based on functionality that switches the theme of the background based on which section that the user is currently in. Going to expand this into the animation.